requirements.txt
flask==3.1.2
twilio==9.8.4
requests==2.32.5
python-dotenv==1.1.1
